% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_knn_cv.R
\name{my_knn_cv}
\alias{my_knn_cv}
\title{K-Nearest Neighbor Cross-validation function}
\usage{
my_knn_cv(train, cl, k_nn, k_cv)
}
\arguments{
\item{train}{input data frame}

\item{cl}{true class value of your training data}

\item{k_nn}{integer representing the number of neighbors}

\item{k_cv}{integer representing the number of folds}
}
\value{
list with following 2 elements:
\code{class}: a vector of the predicted class Yi for all observations
\code{cv_err}: a numeric with the cross-validation mis-classification error
}
\description{
This function performs cross-validation in R
}
\examples{
penguins <- na.omit(my_penguins[, 1:6])
train <- penguins[, 3:6]
my_knn_cv(train, penguins$species, 1, 5)

}
\keyword{classification,}
\keyword{prediction}
