% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_lm.R
\name{my_lm}
\alias{my_lm}
\title{Linear Model function}
\usage{
my_lm(formula, data)
}
\arguments{
\item{formula}{a formula class object}

\item{data}{input data frame}
}
\value{
a table for each coefficient and columns for the Estimate, Std. Error, t-value, and p-value
}
\description{
This function fits a linear model.
}
\examples{
grades_data <- read.csv("https://www.openintro.org/data/csv/gpa.csv")
my_lm(formula = gpa ~ studyweek, data = grades_data)

}
\keyword{covariance}
\keyword{regression,}
