% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_t.test.R
\name{my_t.test}
\alias{my_t.test}
\title{T-test function}
\usage{
my_t.test(x, alternative, mu)
}
\arguments{
\item{x}{a numeric vector}

\item{alternative}{a character string specifying the alternative hypothesis}

\item{mu}{a number indicating the null hypothesis value of the mean}
}
\value{
a list of following elements:
\code{test_stat}: numeric test statistic
\code{df}: degrees of freedom
\code{alternative}: value of the parameter \code{alternative}
\code{p_val}: p-value
}
\description{
This function performs a one sample t-test.
}
\examples{
helium_data <- read.csv("https://www.openintro.org/data/csv/helium.csv")
my_t.test(helium_data[3], alternative = "greater", mu = 20)
}
\keyword{t-test}
